start ship command
